+++ app/src/main/java/org/simple/clinic/selectstate/SelectStateEffect.kt
@@ -7,3 +7,7 @@ sealed class SelectStateEffect
 object LoadStates : SelectStateEffect()
 
 data class SaveSelectedState(val state: State) : SelectStateEffect()
+
+sealed class SelectStateViewEffect : SelectStateEffect()
+
+object GoToRegistrationScreen : SelectStateViewEffect()