+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 - [In Progress: 2 Sep 2021] Add a progress state in `CustomDrugEntrySheet`
 
 ### Features
-- [In-progress: 02 Sep 2021] Add support for state selection after selecting country
+- [In-progress: 06 Sep 2021] Add support for state selection after selecting country
 
 ### Changes
 - Implement providing drug frequencies label depending on the country